feat: basic iTerm image protocol support
This commit implements the functions in issue #175.

Currently, only the iTerm image protocol is supported. Some code from muesli/reflow is copied manually patched, so wrapping and indentation will not interfere with the iTerm escape sequence (which is not quite the same as the ANSI sequences)

This functionality is only enabled if you pass `WithImageDisplay()` when creating a new TermRenderer. I have verified that it works under WezTerm. If the terminal is not supported, it will automatically switch to the old rendering mode of images as fallback.

For more details, please see `TestImageDisplay` in glamour_test.go .